Failing part 2

+++ 2020/runjuliafiles.sh
@@ -6,6 +6,10 @@ function runjulia {
     if [[ "$SUBSTRING" =~ ^Day.* ]]; then
         cp $1* .
         julia $SUBSTRING.jl
+        status=$?
+        if test $status -ne 0
+            exit $status
+        fi
         rm *.jl
         rm *.txt
         # /c/Users/sten.johnsen/AppData/Local/Programs/Julia\ 1.5.3/bin/julia.exe Day21.jl
